+++ R/utils.R
@@ -280,3 +280,25 @@ get_just <- function(element) {
     element$vjust %||% 0.5
   )
 }
+
+.label_params <- setdiff(fn_fmls_names(element_text), c("margin", "debug", "inherit.blank"))
+.line_params <- c("colour", "color", "linewidth", "linetype")
+
+extra_args <- function(..., .valid_args = .label_params, call = caller_env()) {
+  args <- list2(...)
+  if (length(args) == 0) {
+    return(NULL)
+  }
+
+  if (!is.null(args$color)) {
+    args$colour <- args$color
+    args$color <- NULL
+  }
+  extra <- setdiff(names(args), .valid_args)
+  if (length(extra) > 0) {
+    cli::cli_warn("Ignoring unknown parameters: {.and {extra}}.", call = call)
+  }
+  args <- args[lengths(args) > 0]
+  names(args) <- paste0(".", names(args))
+  args
+}