+++ NEWS.md
@@ -1,5 +1,12 @@
 # legendry (development version)
 
+* Added support for dendrograms (#33):
+    * New scale functions `scale_x_dendro()` and `scale_y_dendro()`.
+    * New full guide function: `guide_axis_dendro()`.
+    * New primitive guide function: `primitive_segments()`
+    * New key functions: `key_segment_manual()`, `key_segment_map()` and 
+      `key_dendro()`.
+
 * Fixed bug where `guide_axis_nested(key = key_range_auto(...))` produced 
   duplicated labels (#31)